Merge html and css from old upload page (incomplete)
Since I want to use the CSS reset, I need to re-do the form styles which is
somewhat tricky. I have added rules to restore most of the original look,
except for the form labels and inputs. I need a CSS reference to finish this,
and don't have Internet access ATM.