Replace settings file with template
Since the settings file may contain sensitive information (such as
cryptographic keys), I am removing it from the repo at this juncture (as I am
about to add keys for ReCaptcha integration).

It is replaced with a template file that is a clone of the real settings file,
minus any confidential informaiton. Anybody who wishes to clone the repo and do
work, or run their own installation of the site, will first need to copy
settings_template.cfg to settings.cfg, and fill in the redacted values.